WFLY-20365 - Split OidcWithDeploymentConfigTest for different stability levels
https://issues.redhat.com/browse/WFCORE-7121
